+++ src/main/java/com/udemy/star_d_link/coursereview/Dto/CourseReviewCreateRequestDto.java
@@ -1,4 +1,5 @@
-package com.udemy.star_d_link.Dto;
+package com.udemy.star_d_link.coursereview.Dto;
+
 
 import java.time.LocalDate;
 import java.util.List;
@@ -7,21 +8,22 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Builder
 @Getter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class CourseReviewModifyRequestDto {
+public class CourseReviewCreateRequestDto {
     //private SiteUser siteUser;
     private Integer userId;
     private Integer likeCount;
     private String title;
     private String content;
     private String hashtag;
+    private LocalDate createdAt;
     private LocalDate updatedAt;
     private Integer postType;
     private String name;
     private Integer rating;
-    private List<CourseReviewFileModifyRequestDto> fileDeleteDtoList;
-    private List<CourseReviewFileCreateRequestDto> fileCreateDtoList;
+    private List<CourseReviewFileCreateRequestDto> fileListDto;
+
 }
