+++ src/main/java/com/udemy/star_d_link/coursereview/Dto/CourseReviewFileCreateRequestDto.java
@@ -1,6 +1,7 @@
-package com.udemy.star_d_link.Dto;
+package com.udemy.star_d_link.coursereview.Dto;
 
-import com.udemy.star_d_link.Entity.CourseReview;
+import com.udemy.star_d_link.coursereview.Entity.CourseReview;
+
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,8 +12,10 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class CourseReviewFileModifyRequestDto {
+public class CourseReviewFileCreateRequestDto {
+
     CourseReview courseReview;
     String fileUrl;
+    LocalDate createdAt;
     LocalDate updatedAt;
 }
