+++ src/main/java/com/udemy/star_d_link/coursereview/Dto/CourseReviewFileModifyRequestDto.java
@@ -1,7 +1,6 @@
-package com.udemy.star_d_link.Dto;
+package com.udemy.star_d_link.coursereview.Dto;
 
-import com.udemy.star_d_link.Entity.CourseReview;
-
+import com.udemy.star_d_link.coursereview.Entity.CourseReview;
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -12,10 +11,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class CourseReviewFileCreateRequestDto {
-
+public class CourseReviewFileModifyRequestDto {
     CourseReview courseReview;
     String fileUrl;
-    LocalDate createdAt;
     LocalDate updatedAt;
 }
