+++ src/main/java/com/udemy/star_d_link/coursereview/Dto/CourseReviewListDto.java
@@ -1,4 +1,4 @@
-package com.udemy.star_d_link.Dto;
+package com.udemy.star_d_link.coursereview.Dto;
 
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
