+++ src/main/java/com/udemy/star_d_link/coursereview/Dto/CourseReviewModifyRequestDto.java
@@ -1,5 +1,4 @@
-package com.udemy.star_d_link.Dto;
-
+package com.udemy.star_d_link.coursereview.Dto;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -8,22 +7,21 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
 @Builder
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CourseReviewCreateRequestDto {
+public class CourseReviewModifyRequestDto {
     //private SiteUser siteUser;
     private Integer userId;
     private Integer likeCount;
     private String title;
     private String content;
     private String hashtag;
-    private LocalDate createdAt;
     private LocalDate updatedAt;
     private Integer postType;
     private String name;
     private Integer rating;
-    private List<CourseReviewFileCreateRequestDto> fileListDto;
-
+    private List<CourseReviewFileModifyRequestDto> fileDeleteDtoList;
+    private List<CourseReviewFileCreateRequestDto> fileCreateDtoList;
 }
