+++ src/main/java/com/udemy/star_d_link/coursereview/Repository/CourseReviewFileRepository.java
@@ -1,7 +1,7 @@
-package com.udemy.star_d_link.Repository;
+package com.udemy.star_d_link.coursereview.Repository;
 
-import com.udemy.star_d_link.Entity.CourseReview;
-import com.udemy.star_d_link.Entity.CourseReviewFile;
+import com.udemy.star_d_link.coursereview.Entity.CourseReview;
+import com.udemy.star_d_link.coursereview.Entity.CourseReviewFile;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
